[core] renamed compiler to data extractor
